+++ Projeto Sementes do Brasil/src/br/com/fiap/models/implement/Seguro.java
@@ -1,7 +1,7 @@
 package br.com.fiap.models.implement;
 
 public interface Seguro {
-    int getIdTipoSeguro(); // Retorna o ID do tipo de seguro.
+    long getIdTipoSeguro(); // Retorna o ID do tipo de seguro.
     String getDescricao(); // Retorna a descrição do seguro.
     String getCategoria(); // Retorna a categoria do seguro (Pessoa Física ou Jurídica).
